Cap Nhat

+++ Website/WebContent/threadadmin.jsp
@@ -44,21 +44,21 @@
                            <div class ="panel-body"> 
                   
                          <ul class="nav navs-tabs-brand">
-                                <li class="active"><a href="admin.jsp" class="list-group-item " style="z-index: 0"><i class="glyphicon glyphicon-home"></i> &nbsp;&nbsp;&nbsp;&nbsp;Trang chủ quản trị</a></li>
+                              <li class="active"><a href="quanly-khoa-hoc" class ="list-group-item " style="z-index: 0"><i class="glyphicon glyphicon-education"></i>&nbsp;&nbsp;&nbsp;&nbsp; Quản lý khoá học</a></li>
                                
-                                <li class="active"><a href="quanlykhoahoc.jsp" class ="list-group-item " style="z-index: 0"><i class="glyphicon glyphicon-education"></i>&nbsp;&nbsp;&nbsp;&nbsp; Quản lý khoá học</a></li>
-                               
-                                <li class="active"><a href="threadadmin.jsp" class="list-group-item active" style="z-index: 0"><i class="glyphicon glyphicon-comment"></i> &nbsp;&nbsp;&nbsp;&nbsp;Thread thảo luận</a></li>
+                                <li class="active"><a href="quan-ly-thread" class="list-group-item active" style="z-index: 0"><i class="glyphicon glyphicon-comment"></i> &nbsp;&nbsp;&nbsp;&nbsp;Thread thảo luận</a></li>
                              
-                                <li class="active"><a href="admingiangday.jsp" class ="list-group-item" style="z-index: 0"><i class="glyphicon glyphicon-calendar"></i> &nbsp;&nbsp;&nbsp;&nbsp;Kế hoạch giảng dạy</a></li>
+                                <li class="active"><a href="ke-hoach-giang-day" class ="list-group-item" ><i class="glyphicon glyphicon-calendar"></i> &nbsp;&nbsp;&nbsp;&nbsp;Kế hoạch giảng dạy</a></li>
                               
-                                <li class="active"><a href="adminQuanLyHocVien.jsp" class="list-group-item" style="z-index: 0"><i class="glyphicon glyphicon-user"></i> &nbsp;&nbsp;&nbsp;&nbsp;Quản lý học viên</a></li>
+                                <li class="active"><a href="quan-ly-hoc-vien" class="list-group-item" style="z-index: 0"><i class="glyphicon glyphicon-user"></i> &nbsp;&nbsp;&nbsp;&nbsp;Quản lý học viên</a></li>
                               
-                                <li class="active"><a href="adminGuiMail.jsp" class ="list-group-item" style="z-index: 0"><i class="glyphicon glyphicon-envelope"></i>&nbsp;&nbsp;&nbsp;&nbsp; Gửi mail cho sinh viên</a></li>
-                               <li class="active"><a href="admindanhsachhocvien.jsp" class ="list-group-item" style="z-index: 0"><i class="glyphicon glyphicon-list-alt"></i>&nbsp;&nbsp;&nbsp;&nbsp; Bài tập của học viên</a></li>
+                                <li class="active"><a href="gui-mail" class ="list-group-item" style="z-index: 0"><i class="glyphicon glyphicon-envelope"></i>&nbsp;&nbsp;&nbsp;&nbsp; Gửi mail cho sinh viên</a></li>
+                              
+                            
+                                <li class="active"><a href="danh-sach-hoc-vien-nop-bai" class ="list-group-item" style="z-index: 0"><i class="glyphicon glyphicon-list-alt"></i>&nbsp;&nbsp;&nbsp;&nbsp; Bài tập của học viên</a></li>
                                
-                                <li class="active"><a href="admindethitracnghiem.jsp" class="list-group-item" style="z-index: 0"><i class="glyphicon glyphicon-pencil"></i> &nbsp;&nbsp;&nbsp;&nbsp;Đề thi trắc nghiệm</a></li>
-
+                                <li class="active"><a href="them-de-thi-trac-nghiem" class="list-group-item" style="z-index: 0"><i class="glyphicon glyphicon-pencil"></i> &nbsp;&nbsp;&nbsp;&nbsp;Đề thi trắc nghiệm</a></li>
+                 
                         </ul>
                     </div>  
                     </div>
@@ -93,10 +93,12 @@
                                 				<td>${rows.NgayTao }</td>
                                 				<td>
                                                     <center> 
-                                                     <a href="suathreadadmin.jsp?thread=${rows.MaThread }"> <button type="button" class="btn btn-success" id="btnXacNhan">Sửa</button></a>
-                                                  	 <a href="Thaoluan?command=delete&MaThread=${rows.MaThread }"> <button type="button" class="btn btn-warning" id="btnXoa1">Xóa</button></a> 
+                                                     <a href="chinh-sua-thread?thread=${rows.MaThread }"> <button type="button" class="btn btn-success" id="btnXacNhan" onclick="">Sửa</button></a>
+                                                  	 <a href="Thaoluan?command=delete&MaThread=${rows.MaThread }"> <button type="button"  onclick="return confirmAction()" class="btn btn-warning"  id="btnXoa1" onclick="">Xóa</button></a> 
                                                 	</center>
                                                 </td>
+
+
 
                                 			</tr>
                                 			</c:forEach> 
@@ -104,7 +106,7 @@
                                 	<%-- 	<%} %>
                                 	 --%></table>	
                             
-                              		 <a href="taothreadadmin.jsp"><button type="button" class="btn btn-success" id="btnXacNhan">Tạo thread</button></a>
+                              		 <a href="tao-thread-moi"><button type="button" class="btn btn-success" >Tạo thread</button></a>
                              		
                                </div>
                 
@@ -125,11 +127,13 @@
     <script src="js/bootstrap.min.js"></script>
     
      
-    <script type="text/javascript">
-    function formaction(){
-     return alert('Thao tác thành công');
-        
-    }</script>
+   
+ <SCRIPT LANGUAGE="JavaScript">
+      function confirmAction() {
+        return confirm("Chắc chắn xoá?")
+      }
+ 
+      </SCRIPT>
     
   </body>
 </html>
